refactor(cli): give the win32 shim-name policy one home
appiumBins.ts and playwrightBins.ts held the same 13 lines with one string
changed, including a byte-identical comment. Their tests differed only in the
tool name. spawn.ts tests the resolved name with /\.(cmd|bat)$/i, so every
producer of a shim name has to stay in sync with that regex by hand. One
producer is trackable; two identical copies are not.

appiumCliCandidates and playwrightCliCandidates stay as wrappers over
nodeModulesBinCandidates, so the six consumers changed only their import path.
androidBins.ts stays separate: a different suffix rule over different
directories, not node_modules/.bin.

WIZ-11296
