--- src/core/playwrightBins.test.ts
@@ -1,25 +1,0 @@
-import { describe, expect, it } from "bun:test";
-import { join } from "node:path";
-
-import { playwrightCliCandidates } from "./playwrightBins.js";
-
-describe("playwrightCliCandidates", () => {
-  const envDir = "/proj";
-  const binDir = join(envDir, "node_modules", ".bin");
-
-  it("returns only the extension-less shim on linux and macOS", () => {
-    expect(playwrightCliCandidates(envDir, "linux")).toEqual([
-      join(binDir, "playwright"),
-    ]);
-    expect(playwrightCliCandidates(envDir, "darwin")).toEqual([
-      join(binDir, "playwright"),
-    ]);
-  });
-
-  it("lists the npm .cmd then the bun .exe on win32, and nothing else", () => {
-    expect(playwrightCliCandidates(envDir, "win32")).toEqual([
-      join(binDir, "playwright.cmd"),
-      join(binDir, "playwright.exe"),
-    ]);
-  });
-});
